py_service/requirements.txt
flask
openai-whisper
requests
python-dotenv